Extract and fix query nft dossier
