+++ src/lib.rs
@@ -31,6 +31,7 @@ mod unittest_query_contract_config;
 mod unittest_query_contract_info;
 mod unittest_query_minters;
 mod unittest_query_all_tokens;
+mod unittest_query_nft_dossier;
 mod unittest_query_num_tokens;
 mod unittest_register_receive_nft;
 mod unittest_send;
